backend/requirements.txt
fastapi
uvicorn[standard]
pillow
pycryptodome
python-multipart
